feat: Add predeploy and deploy scripts, update devDependencies
The predeploy and deploy scripts have been added to facilitate the
deployment process to the GitHub Pages. Additionally, the
devDependencies for "@babel/plugin-proposal-private-property-in-object"
and "gh-pages" have been updated to their latest versions. These changes
enable smoother deployment and align the project with the latest
dependencies. Now, the application can be easily deployed to the
specified homepage URL. No associated issues.